a
Eksenler ve açılar kontrol edildi. Accelerometer şimdi çalışıyo.Açılarda düzgün

+++ Source/Common/Inc/calc.h
@@ -23,8 +23,8 @@ extern int8_t Calc_GetByte2sComplement(uint8_t data);
 extern int16_t Calc_GetHalfWord2sComplement(uint16_t data);
 extern int32_t Calc_GetWord2sComplement(uint32_t data);
 
-extern double Calc_GetDiscreteDerivative(float * currVal, float * prevVal, float samplingTime);
-extern double Calc_GetDiscreteIntegral(float * currVal, float * prevSum, float samplingTime);
+extern double Calc_GetDiscreteDerivative(double * currVal, double * prevVal, double samplingTime);
+extern double Calc_GetDiscreteIntegral(double * currVal, double * prevSum, double samplingTime);
 
 
 #endif	/* CALC_H */
